Make operational-context class compatible
