+++ src/OperationalContext/OperationalContext.tsx
@@ -1,20 +1,48 @@
 import * as React from "react"
 
-import { Context, IMessage, MessageType, useOperationalContext, WindowSize } from "./OperationalContext.init"
+export interface WindowSize {
+  width: number
+  height: number
+}
 
-export interface Props {
-  children: (operationalContext: Context) => undefined | React.ReactNode
+export type MessageType = "info" | "success" | "error"
+
+export interface IMessage {
+  body: string
+  type: MessageType
+  onClick?: () => void
+}
+
+export interface Context {
+  pushState?: (url: string) => void
+  replaceState?: (url: string) => void
+  pushMessage: (message: IMessage) => void
+  windowSize: WindowSize
+  loading: boolean
+  setLoading: (isLoading: boolean) => void
 }
 
 /**
- * This component simply wraps OperationalContext in order to allow styleguidist to pick up on
- * it and display it in the documentation page.
+ * Defining a default context value here, used below when instantiating
+ * the context consumer and provider below in order for context to be
+ * correctly detected throughout the application.
  */
-const OperationalContext: React.SFC<Props> = props => {
-  const ctx = useOperationalContext()
-  return <>{props.children({ ...ctx })}</>
+const defaultContext: Context = {
+  pushState: undefined,
+  replaceState: undefined,
+  pushMessage: (_: IMessage) => void 0,
+  windowSize: {
+    width: 1080,
+    height: 640,
+  },
+  loading: false,
+  setLoading: (_: boolean) => void 0,
 }
 
-export default OperationalContext
+const OperationalContext = React.createContext(defaultContext)
 
-export { Context, WindowSize, IMessage, MessageType, useOperationalContext }
+export const { Provider, Consumer } = OperationalContext
+
+export const useOperationalContext = () => React.useContext(OperationalContext)
+
+export default Consumer
