+++ src/OperationalUI/OperationalUI.tsx
@@ -8,8 +8,7 @@ import * as React from "react"
 import ErrorBoundary from "../Internals/ErrorBoundary"
 import Message from "../Internals/Message/Message"
 import Messages from "../Internals/Messages/Messages"
-import { IMessage, MessageType, WindowSize } from "../OperationalContext/OperationalContext"
-import { Provider } from "../OperationalContext/OperationalContext.init"
+import { IMessage, MessageType, Provider, WindowSize } from "../OperationalContext/OperationalContext"
 import Progress from "../Progress/Progress"
 import { darken, DeepPartial } from "../utils"
 import constants, { OperationalStyleConstants } from "../utils/constants"
